requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
apscheduler>=3.10.0
pandas>=2.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
croniter>=2.0.0
email-validator>=2.1.0
